==> notebooks/single-cell/ustekinumab_cohort/02-Tcell_analysis/03-convert_r2py.ipynb <==
working_dir <- "/data/projects/dschaub/ANCA-GN_transcriptomics"
data_dir <- "/data/projects/dschaub/ANCA-GN_transcriptomics/data/single-cell/ustekinumab/integrated_objects"
save_dir <- "/data/projects/dschaub/ANCA-GN_transcriptomics/data/single-cell/ustekinumab/"

setwd(working_dir)
RhpcBLASctl::blas_set_num_threads(1)
library(Seurat)
library(dplyr)
# library(tidyverse)
library(here)
# library(readxl)
# library(Matrix)
library(IMSBSeuratHelpers)
library(ggplot2)
library(MuDataSeurat)
# library(sctransform)
library(future)
library(data.table)
library(xlsx)
plan("multicore", workers = 1)
options(future.globals.maxSize = 20 * 1000 * 1024^2) # 20GB
options(repr.matrix.max.rows = 600, repr.matrix.max.cols = 200)

path <- paste0(data_dir, "/Harmony_Ustekinumab_4PK4PB_T_celltype.rds")
seu <- readRDS(path)

DefaultAssay(seu) <- "RNA"

seu@assays$CITE

str(seu)

seu@assays$RNA@scale.data <- matrix(numeric(), nrow = 0, ncol = 0)
seu@assays$RNA@var.features <- character()
seu@assays$RNA@meta.features <- data.frame(row.names = row.names(seu@assays$RNA@meta.features))
# seu@assays$RNA@key <- "RNA_"
seu@commands <- list()
seu@graphs <- list()
# seu@reductions <- list()
seu@reductions$pca@feature.loadings <- matrix(numeric(), nrow = 0, ncol = 0)
seu@reductions$harmony@feature.loadings <- matrix(numeric(), nrow = 0, ncol = 0)
seu@reductions$harmony@feature.loadings.projected <- matrix(numeric(), nrow = 0, ncol = 0)

path <- paste0(save_dir, "/ANCA_ustekinumab_4PK4PB_Tcells.h5mu")
WriteH5MU(seu, path)
